separate reclaim case from queue
