+++ test/e2e/queue.go
@@ -193,86 +193,4 @@ var _ = Describe("Queue E2E Test", func() {
 			Expect(err).NotTo(HaveOccurred(), "Error waiting for queue unknown")
 		})
 	})
-
-	It("Reclaim", func() {
-		q1, q2 := "reclaim-q1", "reclaim-q2"
-		ctx := initTestContext(options{
-			queues: []string{q1, q2},
-			priorityClasses: map[string]int32{
-				"low-priority":  10,
-				"high-priority": 10000,
-			},
-		})
-		defer cleanupTestContext(ctx)
-
-		slot := oneCPU
-		rep := clusterSize(ctx, slot)
-
-		spec := &jobSpec{
-			tasks: []taskSpec{
-				{
-					img: defaultNginxImage,
-					req: slot,
-					min: 1,
-					rep: rep,
-				},
-			},
-		}
-
-		spec.name = "q1-qj-1"
-		spec.queue = q1
-		spec.pri = "low-priority"
-		job1 := createJob(ctx, spec)
-		err := waitJobReady(ctx, job1)
-		Expect(err).NotTo(HaveOccurred())
-
-		err = waitQueueStatus(func() (bool, error) {
-			queue, err := ctx.vcclient.SchedulingV1beta1().Queues().Get(context.TODO(), q1, metav1.GetOptions{})
-			Expect(err).NotTo(HaveOccurred())
-			return queue.Status.Running == 1, nil
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		expected := int(rep) / 2
-		// Reduce one pod to tolerate decimal fraction.
-		if expected > 1 {
-			expected--
-		} else {
-			err := fmt.Errorf("expected replica <%d> is too small", expected)
-			Expect(err).NotTo(HaveOccurred())
-		}
-
-		spec.name = "q2-qj-2"
-		spec.queue = q2
-		spec.pri = "high-priority"
-		job2 := createJob(ctx, spec)
-		err = waitTasksReady(ctx, job2, expected)
-		Expect(err).NotTo(HaveOccurred())
-
-		err = waitTasksReady(ctx, job1, expected)
-		Expect(err).NotTo(HaveOccurred())
-
-		// Test Queue status
-		spec = &jobSpec{
-			name:  "q1-qj-2",
-			queue: q1,
-			tasks: []taskSpec{
-				{
-					img: defaultNginxImage,
-					req: slot,
-					min: rep * 2,
-					rep: rep * 2,
-				},
-			},
-		}
-		job3 := createJob(ctx, spec)
-		err = waitJobStatePending(ctx, job3)
-		Expect(err).NotTo(HaveOccurred())
-		err = waitQueueStatus(func() (bool, error) {
-			queue, err := ctx.vcclient.SchedulingV1beta1().Queues().Get(context.TODO(), q1, metav1.GetOptions{})
-			Expect(err).NotTo(HaveOccurred())
-			return queue.Status.Pending == 1, nil
-		})
-		Expect(err).NotTo(HaveOccurred())
-	})
 })
